refactor(config): split config.go into 14 domain-specific files
Split config.go (1638 lines, 32 structs) into focused files by domain:
- config.go: Config struct, Load(), Validate(), small structs (CORS, Admin, etc.)
- config_server.go: ServerConfig, BodyLimitsConfig
- config_database.go: DatabaseConfig + connection string methods
- config_auth.go: AuthConfig, SAMLProviderConfig, OAuthProviderConfig
- config_security.go: SecurityConfig, CaptchaConfig, AdaptiveTrustConfig
- config_storage.go: StorageConfig, TransformConfig
- config_email.go: EmailConfig
- config_functions.go: FunctionsConfig
- config_api.go: APIConfig
- config_jobs.go: JobsConfig
- config_ai.go: AIConfig
- config_telemetry.go: TracingConfig, MetricsConfig, LoggingConfig
- config_tenants.go: TenantsConfig + sub-types
- config_scaling.go: ScalingConfig
